server/requirements.txt
Django==4.2.4
django_storages[google]==1.13.2
kubernetes==27.2.0
graphene-django==3.1.5
django-filter==23.2
specklepy==2.16.0
graphene-file-upload==1.3.0
python-slugify==8.0.1
django-cors-headers==4.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.6
google-cloud-logging==3.6.0